feat: wait till server fully initialised
This change makes sure a server node has either connected to an existing
cluster or started a new one before initialisation completes. This is
necessary to prevent writes before the cluster has formed - although this
is largely because the connection logic is conservative atm and does not
attempt to merge even orthogonal partitions.

The test mimics two servers which update at the earliest opportunity,
then one node dies, and the replacement picks up smoothly from before.